rename extract to export, for familiarity to next.js devs
